qemu-binaries: Initial commit
The point of this repo is to be able to build versions of QEMU for
validating Clang built kernels. Currently, we only require a copy of
qemu-system-s390x, due to the series linked in the README.

We use docker or podman to build the image so that we can run it through
GitHub Actions. It is a static binary but there are some parts of it
that need the same glibc version that it was linked against.

It is compressed with zstd so that this repo does not get too large in
size.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>